+++ Dependencies/Renderer.Wpf/WpfRenderer.cs
@@ -118,10 +118,10 @@ namespace Dependencies
         {
             var brush = new SolidColorBrush(
                 Color.FromArgb(
-                    (byte)color.A,
-                    (byte)color.R,
-                    (byte)color.G,
-                    (byte)color.B));
+                    color.A,
+                    color.R,
+                    color.G,
+                    color.B));
             brush.Freeze();
             return brush;
         }
